:building_construction: Move data to new data folder
